+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the Arduino Maker Workshop extension will be documented i
 
 - Added a Home page warning and Open Folder action when VS Code has no workspace folder open
 - Added Sketchbook folder controls to Library Manager so users can choose where Arduino libraries are installed ([issue #112](https://github.com/thelastoutpostworkshop/arduino-maker-workshop/issues/112))
+- Preserved user C/C++ configurations when regenerating IntelliSense settings and added build profile build properties ([issue #109](https://github.com/thelastoutpostworkshop/arduino-maker-workshop/issues/109))
 
 ## Version 1.1.4
 - Added a setting to disable automatic board/library update checks for offline usage ([issue #108](https://github.com/thelastoutpostworkshop/arduino-maker-workshop/issues/108))